Split PWM enable from setting compare value
This allows more flexibility in choosing the brightness of blinking,
currently used only to flicker at the current brightness for battery
warning.

Sadly due to additional calls to set low intensity for config mode this
saves only 2 bytes of flash memory, but with some other parameters it is
possible to save up to 6 bytes more.