There's something strange...
... in the event handling.

This version produces some weird results. Sometimes a single press
in the strobe mode switches to the next strobe pattern while in other
times it switches to a fixed mode.

Even more confusingly there is a new "mode": Ramping up and down and
slow double flash at low.

I assume that the compiler made assumptions about the flow that aren't
true. But pressing the button should start from the beginning and the
only independent variable that can cause this issue is (or should be)
state. Strange...