Fix #2: Add strobe mode
If users want a strobe mode, they usually want it for defence purposes.
Thus it should be possible to enter it as fast as possible. If strobe is
enabled (first option if STROBE is compiled) the flashlight always
starts in a strobe mode. Then a single press of the power switch enters
the usual UI.