Fix #6: Moonlight mode
Extends the strobe mode. If options.moonlight is set and options.strobe
is not, then the light will turn on alway in moonlight and after a
single tap it will continue with the default mode.

At least with the development board there is no visibly huge difference
in output, so this should be considered WIP.